[6/n][a2][writer] add writer component
Summary:

refer to design hierarchy.

we want Paxos class to be composed of different components. naturally we have:
- proposer
- accepter

Each need a reader and a writer component to interface with underlying GCL

this diff provides the primitive implementation for the reader component

Test Plan:

builds and runs at top of stack