[7/n][a2][reader] add reader component
Summary:

refer to previous diff for more context. this diff implements the reader component.

reader maintains a queue for proposer to ingest from and another for acceptor to ingest from.

why? because GCL only submits 1 message. if we had 2 readers (1 for proposer, 1 for acceptor), then they might end up 'fighting' for msgs. consider the case where a message is sent and reader1 picks up the msg before reader2. since there is now no more messages coming in, reader2 is blocked and the component holding reader2 is halted.

we could implement a message duplicator to read from the GCL and duplicate each message, but I leave that up to the reviewer to decide.

Test Plan:

it builds and runs at top of stack