[3/n][a3][client] refactor client
Summary:

ditch the idea of creating separate class files for each callback/watcher...

project size isn't big enough to justify that, just keep private lambda definitions on hand within the class and call those.

this way, if you need to edit a file, you don't need to go opening every other callback/watcher file and have everything on hand. Up to reviewer to decide if they want these private lambdas to be put at the top or bottom of file (or if should be done another way).

Test Plan:

ig it builds so that's a good sign right